+++ arm9/source/gdi.h
@@ -51,11 +51,11 @@ class cGdi {
     void maskBlt(const void* src, s16 destX, s16 destY, u16 destW, u16 destH, GRAPHICS_ENGINE engine);
     void maskBlt(const void* src, s16 destX, s16 destY, u16 destW, u16 destH, GRAPHICS_ENGINE engine, u16 color);
     void maskBlt(const void* src, s16 srcW, s16 srcH, s16 destX, s16 destY, u16 destW, u16 destH, GRAPHICS_ENGINE engine);
-    void maskBlt(const void* src, s16 srcW, s16 srcH, s16 destX, s16 destY, u16 destW, u16 destH, GRAPHICS_ENGINE engine, u16 color);
+    void maskBlt(const void* src, s16 srcW, s16 srcH, s16 destX, s16 destY, s32 destW, s32 destH, GRAPHICS_ENGINE engine, u16 color);
     void bitBlt(const void* src, s16 destX, s16 destY, u16 destW, u16 destH, GRAPHICS_ENGINE engine);
     void bitBlt(const void* src, s16 destX, s16 destY, u16 destW, u16 destH, u16 repeats, GRAPHICS_ENGINE engine);
     void bitBlt(const void* src, s16 srcW, s16 srcH, s16 destX, s16 destY, u16 destW, u16 destH, GRAPHICS_ENGINE engine);
-    void bitBlt(const void* src, s16 srcW, s16 srcH, s16 destX, s16 destY, u16 destW, u16 destH, u16 repeats, GRAPHICS_ENGINE engine);
+    void bitBlt(const void* src, s16 srcW, s16 srcH, s16 destX, s16 destY, s32 destW, s32 destH, u16 repeats, GRAPHICS_ENGINE engine);
 
     void bitMainBackground(const void* src);
     void bitSubBackground(const void* src);
